+++ tools/resourcemanager/src/main/java/org/sgdk/resourcemanager/entities/SGDKElement.java
@@ -52,7 +52,10 @@ public abstract class SGDKElement {
 			throw e;
 		}
 		if(!validateFormat(path)) throw new SGDKInvalidFormatException();
+		init();
 	}
+	
+	protected abstract void init() throws SGDKInvalidFormatException;
 	
 	public String getPath() {
 		return path;