+++ tools/resourcemanager/src/main/java/org/sgdk/resourcemanager/entities/SGDKProject.java
@@ -17,12 +17,10 @@ public class SGDKProject extends SGDKFolder{
 	
 	public SGDKProject(JsonNode node) throws SGDKInvalidFormatException, IOException {
 		super(node);
-		setType(Type.SGDKProject);
 	};
 	
 	public SGDKProject(String path) throws SGDKInvalidFormatException, IOException {
 		super(path);
-		setType(Type.SGDKProject);
 	}
 	
 	@Override
@@ -31,5 +29,11 @@ public class SGDKProject extends SGDKFolder{
 				getClass().getResource("/icons/053-folder-24.svg").toURI(),
 				16,
 				16);
+	}
+	
+	@Override
+	protected void init() throws SGDKInvalidFormatException {
+		super.init();
+		setType(Type.SGDKProject);
 	}
 }